moved timing function to its own file
