+++ examples/low/nano.h
@@ -1,0 +1,8 @@
+#ifndef NANO_H
+#define NANO_H
+
+#include <time.h>
+
+double nano(struct timespec *start, struct timespec *end);
+
+#endif
